+++ composeApp/src/commonMain/kotlin/org/wip/plugintoolkit/features/settings/viewmodel/SettingsViewModel.kt
@@ -156,4 +156,8 @@ class SettingsViewModel(
     fun openLogFolder() {
         repository.openLogFolder()
     }
+
+    fun openLatestLog() {
+        repository.openLatestLog()
+    }
 }